refactor: 공통 Infra 패키지 생성
