+++ src/main/java/com/example/sns/common/infrastructure/imagestore/ImageStore.java
@@ -1,9 +1,7 @@
-package com.example.sns.common.imagestore;
+package com.example.sns.common.infrastructure.imagestore;
 
-import com.example.sns.common.imagestore.exception.ImageStoreException;
+import com.example.sns.common.infrastructure.imagestore.exception.ImageStoreException;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
 
 public interface ImageStore {
     String saveImage(MultipartFile image, String storePath) throws ImageStoreException;
