+++ src/main/java/com/example/sns/common/infrastructure/imagestore/exception/TemporaryFileException.java
@@ -1,4 +1,4 @@
-package com.example.sns.common.imagestore.exception;
+package com.example.sns.common.infrastructure.imagestore.exception;
 
 import org.springframework.http.HttpStatus;
